+++ src/main.cpp
@@ -2,7 +2,7 @@
 #include <unordered_set>
 #include <pthread.h>
 #include "linked_list.h"
-#include "threads.h"
+#include "threading.h"
 
 void print_usage();
